+++ src/Controller/Admin/OrderArticle.php
@@ -10,6 +10,7 @@ namespace OxidSolutionCatalysts\AmazonPay\Controller\Admin;
 use OxidEsales\Eshop\Core\Exception\DatabaseConnectionException;
 use OxidEsales\Eshop\Core\Exception\DatabaseErrorException;
 use OxidEsales\EshopCommunity\Core\Request;
+use OxidSolutionCatalysts\AmazonPay\Core\Config;
 use OxidSolutionCatalysts\AmazonPay\Core\Constants;
 use OxidSolutionCatalysts\AmazonPay\Core\Logger;
 use OxidSolutionCatalysts\AmazonPay\Core\Provider\OxidServiceProvider;
@@ -35,6 +36,10 @@ class OrderArticle extends OrderArticle_parent
      */
     private function refundAmazon()
     {
+        $config = new Config();
+        if (!$config->automatedRefundActivated()) {
+            return;
+        }
         $request = new Request();
         // get article id
         /** @var string $sOrderArtId */
@@ -54,7 +59,7 @@ class OrderArticle extends OrderArticle_parent
                 $logger = new Logger();
                 OxidServiceProvider::getAmazonService()->createRefund(
                     $oOrder->getId(),
-                    (float)$oOrderArticle->getTotalBrutPriceFormated(),
+                    (float)$oOrderArticle->getFieldData('oxbrutprice'),
                     $logger
                 );
             }