+++ src/Core/Config.php
@@ -17,7 +17,7 @@ use OxidEsales\Eshop\Core\Exception\DatabaseConnectionException;
 use OxidEsales\Eshop\Core\Exception\DatabaseErrorException;
 use OxidEsales\Eshop\Core\Exception\StandardException;
 use OxidEsales\Eshop\Core\Registry;
-use OxidEsales\EshopCommunity\Internal\Container\ContainerFactory;
+use OxidEsales\EshopCommunity\Application\Model\CountryList;
 use OxidSolutionCatalysts\AmazonPay\Core\Provider\OxidServiceProvider;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\Bridge\ModuleSettingBridgeInterface;
 
@@ -415,6 +415,32 @@ class Config
     /**
      * @return bool
      */
+    public function automatedRefundActivated(): bool
+    {
+        $moduleSettingBridge = ContainerFactory::getInstance()
+            ->getContainer()
+            ->get(ModuleSettingBridgeInterface::class);
+        /** @var string $blAmazonAutomatedRefundActivated */
+        $blAmazonAutomatedRefundActivated = $moduleSettingBridge->get('blAmazonAutomatedRefundActivated', AmazonPayModule::MODULE_ID);
+        return (bool) $blAmazonAutomatedRefundActivated;
+    }
+
+    /**
+     * @return bool
+     */
+    public function automatedCancelActivated(): bool
+    {
+        $moduleSettingBridge = ContainerFactory::getInstance()
+            ->getContainer()
+            ->get(ModuleSettingBridgeInterface::class);
+        /** @var string $blAmazonAutomatedCancelActivated */
+        $blAmazonAutomatedCancelActivated = $moduleSettingBridge->get('blAmazonAutomatedCancelActivated', AmazonPayModule::MODULE_ID);
+        return (bool) $blAmazonAutomatedCancelActivated;
+    }
+
+    /**
+     * @return bool
+     */
     public function displayExpressInMiniCartAndModal(): bool
     {
         $moduleSettingBridge = ContainerFactory::getInstance()
@@ -502,7 +528,17 @@ class Config
             $payment = oxNew(Payment::class);
             /** TODO should be variable for any amazonpay ID */
             $payment->load(Constants::PAYMENT_ID_EXPRESS);
-            foreach ($payment->getCountries() as $countryOxId) {
+            $allowedCountries = $payment->getCountries();
+            // fallback if countries are not restricted by Paymentmethod ...
+            if (!$allowedCountries) {
+                $allowedCountries = [];
+                $countries = oxNew(CountryList::class);
+                $countries->loadActiveCountries();
+                foreach ($countries as $allowedCountry) {
+                    $allowedCountries[] = $allowedCountry->getId();
+                }
+            }
+            foreach ($allowedCountries as $countryOxId) {
                 // check deliverysets
                 $deliverySetList = oxNew(DeliverySetList::class);
                 $deliverySetData = $deliverySetList->getDeliverySetList($activeUser, $countryOxId);