+++ src/Core/Constants.php
@@ -84,6 +84,11 @@ class Constants
         return !empty(self::PAYMENT_DESCRIPTIONS[$paymentId]);
     }
 
+    public static function isAmazonExpressPayment(string $paymentId): bool
+    {
+        return ($paymentId === self::PAYMENT_ID_EXPRESS && self::isAmazonPayment($paymentId));
+    }
+
     public static function getPaymentIds(): array
     {
         return array_keys(self::PAYMENT_DESCRIPTIONS);
@@ -96,21 +101,21 @@ class Constants
         self::PAYMENT_ID => [
             'en' => [
                 'title' => 'AmazonPay',
-                'desc' => 'AmazonPay'
+                'desc' => '<div>AmazonPay</div>'
             ],
             'de' => [
                 'title' => 'AmazonPay',
-                'desc' => 'AmazonPay'
+                'desc' => '<div>AmazonPay</div>'
             ]
         ],
         self::PAYMENT_ID_EXPRESS => [
             'en' => [
                 'title' => 'AmazonPay Express',
-                'desc' => 'AmazonPay Express'
+                'desc' => '<div>AmazonPay Express</div>'
             ],
             'de' => [
                 'title' => 'AmazonPay Express',
-                'desc' => 'AmazonPay Express'
+                'desc' => '<div>AmazonPay Express</div>'
             ]
         ],
     ];