+++ CHANGELOG.md
@@ -8,6 +8,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 * Delivery address was not send to Amazon
+* Fixed IPN history list in adminarea
+* Change requirement for PHP7.0 compatibility
+* In addition to the country restrictions on the payment method, also check the country restrictions of the shop.
+* Send orderNr as ReferenceMerchantId to Amazon
+* calculate AmazonPayExpress-Deliverycosts based on provided Country from Amazon
+* [0007379](https://bugs.oxid-esales.com/view.php?id=7379) Fix Error messages from the DispatchController spam the log
+* Do not duplicate IPN and transaction history entries in order backend
+* [0007508](https://bugs.oxid-esales.com/view.php?id=7508) now automated refund and cancel are optional
+* [0007501](https://bugs.oxid-esales.com/view.php?id=7501) fix to buy variants in OXID
+* Show Charge-Status additionally from API in Admin-Order-Overview
+* Config-Values still exists after (de-)activating in OXID >=6.3
 
 ## [2.1.2] - Release
 
@@ -25,12 +36,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * [0007471](https://bugs.oxid-esales.com/view.php?id=7471) Module throws an exception in case the button Sign off is used during the checkout
 * [0007455](https://bugs.oxid-esales.com/view.php?id=7455) if Amazon Pay module is active, then the shipping address is ignored for all payment methods
 * [0007462](https://bugs.oxid-esales.com/view.php?id=7462) if mandatory Confirm General Terms and Conditions field is activated -> you go round in circles
-* compatibility issues with core or other payment-modules 
+* compatibility issues with core or other payment-modules
 
 ## [2.1.0] - Release
 
 ### Added
- 
+
 * added additional payment button
 * If the order or individual items are canceled or deleted, Amazon will issue a refund or cancel
 * Added log in via Amazon
@@ -43,7 +54,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * [0007463](https://bugs.oxid-esales.com/view.php?id=7463) reAdd missing MerchantReferenceId
 * [0007350](https://bugs.oxid-esales.com/view.php?id=7350) OXTRANSID contains authorize status codes
 
-## [2.0.1] - Release 
+## [2.0.1] - Release
 
 ### Added